Separate gameplay helpers from agent dependencies
Move lifecycle hooks, screenshot encoding, battle predicates, and facing-tile lookup to their owning modules. Update callers and remove unnecessary import aliases without changing gameplay behavior.
