+++ agent/battle/tools/throw_ball/service.py
@@ -3,7 +3,7 @@
 from typing import TYPE_CHECKING
 
 from agent.battle.tools.errors import BattleActionUnavailableError
-from agent.battle.tools.utils import get_cursor_pos_in_fight_menu
+from agent.battle.utils import get_cursor_pos_in_fight_menu
 from agent.utils import move_cursor
 from common.enums import BattleType, Button, PokeballItem
