+++ agent/overworld/agent.py
@@ -8,14 +8,12 @@ from pydantic_ai.models.openai import OpenAIResponsesModelSettings
 from pydantic_graph import End
 
 from agent.context import AgentContext
+from agent.formatting.game_state import build_screenshot_content
+from agent.hooks import AGENT_HOOKS
 from agent.overworld.map_view import CurrentMapView, build_current_map_view
 from agent.overworld.prompts import build_overworld_decision_prompt
 from agent.overworld.tools.registry import build_overworld_toolset
-from agent.utils import (
-    AGENT_HOOKS,
-    build_screenshot_content,
-    is_overworld_handler_state,
-)
+from agent.utils import is_overworld_handler_state
 from common.prompts import SYSTEM_PROMPT
 from llm.service import MODEL, REASONING_EFFORT, TIMEOUT_SECONDS
 from overworld_map.service import prepare_overworld_map
